[TASK-202] Extract shared transcript/pricing utilities into tusk-pricing-lib.py
