+++ bin/tusk-session-stats.py
@@ -14,248 +14,26 @@ Arguments received from tusk:
     sys.argv[3:] — session_id + optional transcript path
 """
 
-import json
+import importlib.util
 import logging
 import os
 import sqlite3
 import sys
-from datetime import datetime, timezone
 from pathlib import Path
 
 log = logging.getLogger(__name__)
 
-# Per-million-token pricing (USD) and model aliases are loaded from
-# pricing.json at runtime. See load_pricing() below.
-PRICING: dict = {}
-MODEL_ALIASES: dict = {}
+
+def _load_lib():
+    """Import tusk-pricing-lib.py (hyphenated filename requires importlib)."""
+    lib_path = Path(__file__).resolve().parent / "tusk-pricing-lib.py"
+    spec = importlib.util.spec_from_file_location("tusk_pricing_lib", lib_path)
+    mod = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(mod)
+    return mod
 
 
-def load_pricing() -> None:
-    """Load model pricing and aliases from pricing.json.
-
-    Searches next to this script first (installed layout), then the
-    parent directory (source repo layout where pricing.json is at the root).
-    """
-    global PRICING, MODEL_ALIASES
-    script_dir = Path(__file__).resolve().parent
-    candidates = [
-        script_dir / "pricing.json",
-        script_dir.parent / "pricing.json",
-    ]
-    for path in candidates:
-        if path.is_file():
-            log.debug("Loading pricing from %s", path)
-            with open(path) as f:
-                data = json.load(f)
-            PRICING = data.get("models", {})
-            MODEL_ALIASES = data.get("aliases", {})
-            return
-    print(
-        f"Warning: pricing.json not found (searched {', '.join(str(p) for p in candidates)}). "
-        "Cost calculations will return $0.",
-        file=sys.stderr,
-    )
-
-
-def resolve_model(model_id: str) -> str:
-    """Normalize a model ID to a canonical pricing key."""
-    if model_id in PRICING:
-        return model_id
-    if model_id in MODEL_ALIASES:
-        resolved = MODEL_ALIASES[model_id]
-        log.debug("Model alias: %s -> %s", model_id, resolved)
-        return resolved
-    # Try stripping date suffix (e.g. "claude-opus-4-6-20260101")
-    for key in PRICING:
-        if model_id.startswith(key):
-            log.debug("Model prefix match: %s -> %s", model_id, key)
-            return key
-    log.debug("Unknown model (no pricing): %s", model_id)
-    return model_id
-
-
-def find_transcript(project_dir: str) -> str | None:
-    """Find the most recently modified JSONL in the Claude projects dir."""
-    claude_dir = Path.home() / ".claude" / "projects" / project_dir
-    log.debug("Looking for transcripts in %s", claude_dir)
-    if not claude_dir.is_dir():
-        log.debug("Directory does not exist")
-        return None
-    jsonl_files = list(claude_dir.glob("*.jsonl"))
-    log.debug("Found %d JSONL files", len(jsonl_files))
-    if not jsonl_files:
-        return None
-    chosen = str(max(jsonl_files, key=lambda p: p.stat().st_mtime))
-    log.debug("Selected transcript: %s", chosen)
-    return chosen
-
-
-def derive_project_hash(cwd: str) -> str:
-    """Derive Claude Code's project hash from a directory path.
-
-    Claude Code uses the absolute path with '/' replaced by '-',
-    e.g. /Users/foo/myproject -> -Users-foo-myproject
-    """
-    return cwd.replace("/", "-")
-
-
-def parse_timestamp(ts: str) -> datetime:
-    """Parse an ISO 8601 timestamp, handling both Z and +00:00 suffixes."""
-    ts = ts.replace("Z", "+00:00")
-    return datetime.fromisoformat(ts)
-
-
-def parse_sqlite_timestamp(ts: str) -> datetime:
-    """Parse a SQLite datetime string (UTC, no timezone info)."""
-    return datetime.strptime(ts, "%Y-%m-%d %H:%M:%S").replace(tzinfo=timezone.utc)
-
-
-def aggregate_session(
-    transcript_path: str,
-    started_at: datetime,
-    ended_at: datetime | None,
-) -> dict:
-    """Parse a JSONL transcript and aggregate tokens within the time window.
-
-    Returns dict with keys: input_tokens, output_tokens,
-    cache_creation_input_tokens, cache_read_input_tokens, model, request_count.
-    """
-    log.debug("Aggregating session from %s", transcript_path)
-    log.debug("Time window: %s .. %s", started_at.isoformat(),
-              ended_at.isoformat() if ended_at else "now")
-    seen_requests: set[str] = set()
-    totals = {
-        "input_tokens": 0,
-        "output_tokens": 0,
-        "cache_creation_input_tokens": 0,
-        "cache_creation_5m_tokens": 0,
-        "cache_creation_1h_tokens": 0,
-        "cache_read_input_tokens": 0,
-    }
-    model_counts: dict[str, int] = {}
-    request_count = 0
-    lines_read = 0
-
-    with open(transcript_path) as f:
-        for line in f:
-            lines_read += 1
-            line = line.strip()
-            if not line:
-                continue
-            try:
-                entry = json.loads(line)
-            except json.JSONDecodeError:
-                continue
-
-            # Only assistant messages have usage data
-            if entry.get("type") != "assistant":
-                continue
-
-            # Check timestamp is within session window
-            ts_str = entry.get("timestamp")
-            if not ts_str:
-                continue
-            try:
-                ts = parse_timestamp(ts_str)
-            except (ValueError, TypeError):
-                continue
-
-            if ts < started_at:
-                continue
-            if ended_at and ts > ended_at:
-                continue
-
-            # Deduplicate by requestId (streaming produces multiple entries)
-            request_id = entry.get("requestId")
-            if not request_id:
-                continue
-            if request_id in seen_requests:
-                continue
-            seen_requests.add(request_id)
-            request_count += 1
-
-            # Extract usage
-            message = entry.get("message", {})
-            usage = message.get("usage", {})
-            if not usage:
-                continue
-
-            totals["input_tokens"] += usage.get("input_tokens", 0)
-            totals["output_tokens"] += usage.get("output_tokens", 0)
-            totals["cache_read_input_tokens"] += usage.get(
-                "cache_read_input_tokens", 0
-            )
-
-            # Per-tier cache write tokens: prefer the nested cache_creation
-            # object (ephemeral_5m_input_tokens / ephemeral_1h_input_tokens).
-            # Fall back to assigning all cache_creation_input_tokens to the
-            # 5m tier when the nested object is absent (older transcripts).
-            cache_creation = usage.get("cache_creation")
-            cache_total = usage.get("cache_creation_input_tokens", 0)
-            if isinstance(cache_creation, dict):
-                tokens_5m = cache_creation.get("ephemeral_5m_input_tokens", 0)
-                tokens_1h = cache_creation.get("ephemeral_1h_input_tokens", 0)
-                totals["cache_creation_5m_tokens"] += tokens_5m
-                totals["cache_creation_1h_tokens"] += tokens_1h
-            else:
-                totals["cache_creation_5m_tokens"] += cache_total
-            totals["cache_creation_input_tokens"] += cache_total
-
-            # Track model usage
-            model = message.get("model", "")
-            if model:
-                model = resolve_model(model)
-                model_counts[model] = model_counts.get(model, 0) + 1
-
-    log.debug("Lines read: %d, unique requests: %d, duplicates skipped: %d",
-              lines_read, request_count, len(seen_requests) - request_count
-              if len(seen_requests) > request_count else 0)
-    log.debug("Token totals: %s", totals)
-    log.debug("Model counts: %s", model_counts)
-
-    # Determine dominant model
-    dominant_model = ""
-    if model_counts:
-        dominant_model = max(model_counts, key=model_counts.get)
-    log.debug("Dominant model: %s", dominant_model)
-
-    return {
-        **totals,
-        "model": dominant_model,
-        "model_counts": model_counts,
-        "request_count": request_count,
-    }
-
-
-def compute_cost(totals: dict) -> float:
-    """Compute cost in dollars from token totals and model.
-
-    Uses five terms: input, cache_write_5m, cache_write_1h, cache_read, output.
-    """
-    model = totals.get("model", "")
-    rates = PRICING.get(model)
-    if not rates:
-        log.debug("No pricing for model %r — cost = $0", model)
-        return 0.0
-
-    mtok = 1_000_000
-    cost = (
-        totals["input_tokens"] / mtok * rates["input"]
-        + totals["cache_creation_5m_tokens"] / mtok * rates["cache_write_5m"]
-        + totals["cache_creation_1h_tokens"] / mtok * rates["cache_write_1h"]
-        + totals["cache_read_input_tokens"] / mtok * rates["cache_read"]
-        + totals["output_tokens"] / mtok * rates["output"]
-    )
-    log.debug("Cost breakdown (model=%s): input=%d*$%.2f + cache_write_5m=%d*$%.2f "
-              "+ cache_write_1h=%d*$%.2f + cache_read=%d*$%.2f + output=%d*$%.2f = $%.6f",
-              model,
-              totals["input_tokens"], rates["input"],
-              totals["cache_creation_5m_tokens"], rates["cache_write_5m"],
-              totals["cache_creation_1h_tokens"], rates["cache_write_1h"],
-              totals["cache_read_input_tokens"], rates["cache_read"],
-              totals["output_tokens"], rates["output"],
-              cost)
-    return round(cost, 6)
+lib = _load_lib()
 
 
 def main():
@@ -271,7 +49,7 @@ def main():
         stream=sys.stderr,
     )
 
-    load_pricing()
+    lib.load_pricing()
 
     if len(argv) < 3:
         print(
@@ -307,14 +85,14 @@ def main():
         conn.close()
         sys.exit(1)
 
-    started_at = parse_sqlite_timestamp(row["started_at"])
-    ended_at = parse_sqlite_timestamp(row["ended_at"]) if row["ended_at"] else None
+    started_at = lib.parse_sqlite_timestamp(row["started_at"])
+    ended_at = lib.parse_sqlite_timestamp(row["ended_at"]) if row["ended_at"] else None
 
     # Discover transcript if not provided
     if not transcript_path:
         cwd = os.getcwd()
-        project_hash = derive_project_hash(cwd)
-        transcript_path = find_transcript(project_hash)
+        project_hash = lib.derive_project_hash(cwd)
+        transcript_path = lib.find_transcript(project_hash)
         if not transcript_path:
             print(
                 f"Error: No JSONL transcripts found for project hash '{project_hash}'.\n"
@@ -331,7 +109,7 @@ def main():
         sys.exit(1)
 
     # Aggregate tokens
-    totals = aggregate_session(transcript_path, started_at, ended_at)
+    totals = lib.aggregate_session(transcript_path, started_at, ended_at)
 
     if totals["request_count"] == 0:
         print(
@@ -342,13 +120,9 @@ def main():
         conn.close()
         sys.exit(0)
 
-    tokens_in = (
-        totals["input_tokens"]
-        + totals["cache_creation_input_tokens"]
-        + totals["cache_read_input_tokens"]
-    )
+    tokens_in = lib.compute_tokens_in(totals)
     tokens_out = totals["output_tokens"]
-    cost = compute_cost(totals)
+    cost = lib.compute_cost(totals)
     model = totals["model"]
 
     # Update DB
